pyproject.toml
[project]
name = "g-news-summarization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "backcall==0.2.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "chromadb==1.0.12",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "comm==0.2.2",
    "cssselect==1.3.0",
    "curl-cffi==0.11.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docopt==0.6.2",
    "durationpy==0.10",
    "email-validator==2.2.0",
    "executing==2.2.0",
    "fastapi==0.115.9",
    "fastapi-cli==0.0.7",
    "fastjsonschema==2.21.1",
    "feedfinder2==0.0.4",
    "feedparser==6.0.11",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "google-api-core==2.24.2",
    "google-auth==2.40.0",
    "google-genai==1.13.0",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.1",
    "grpcio==1.73.0",
    "h11==0.16.0",
    "hf-xet==1.1.3",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.33.0",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==6.29.5",
    "ipython==8.12.3",
    "jedi==0.19.2",
    "jieba3k==0.35.1",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyterlab-pygments==0.3.0",
    "kubernetes==33.1.0",
    "lxml==5.4.0",
    "lxml-html-clean==0.4.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.1.3",
    "mmh3==5.1.0",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "newspaper3k==0.2.8",
    "nltk==3.9.1",
    "numpy==2.2.5",
    "oauthlib==3.2.2",
    "ollama==0.4.8",
    "onnxruntime==1.22.0",
    "opentelemetry-api==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-instrumentation==0.55b1",
    "opentelemetry-instrumentation-asgi==0.55b1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "opentelemetry-util-http==0.55b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "playwright==1.52.0",
    "posthog==4.7.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pytest>=8.4.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-file==2.1.0",
    "requests-oauthlib==2.0.0",
    "responses==0.24.1",
    "rich==14.0.0",
    "rich-toolkit==0.14.5",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "sgmllib3k==1.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "starlette==0.45.3",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tinycss2==1.4.0",
    "tinysegmenter==0.3",
    "tldextract==5.3.0",
    "tokenizers==0.21.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.15.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarg==0.1.9",
    "yarl==1.20.0",
    "zipp==3.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]
